+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Bug Fixes
+
+- The `sentry.origin` log event attribute is now correctly prefixed with `auto.log` ([#2749](https://github.com/getsentry/sentry-ruby/pull/2749))
+
 ## 5.28.0
 
 ### Features
@@ -7,7 +13,7 @@
 ### Miscellaneous
 
 - Deprecate all Metrics related APIs [#2726](https://github.com/getsentry/sentry-ruby/pull/2726)
-    
+
     Sentry [no longer has the Metrics Beta offering](https://sentry.zendesk.com/hc/en-us/articles/26369339769883-Metrics-Beta-Ended-on-October-7th) so
     all the following APIs linked to Metrics have been deprecated and will be removed in the next major.